+++ example/lib/page_3d_audio.dart
@@ -105,7 +105,6 @@ class _Page3DAudioState extends State<Page3DAudio> {
     final playRet = await SoLoud.instance.play3d(currentSound!, 0, 0, 0);
     if (playRet.error != PlayerErrors.noError) return;
 
-    SoLoud.instance.setLooping(playRet.newHandle, true);
     currentSound = playRet.sound;
 
     spinAround = true;
@@ -130,10 +129,15 @@ class _Page3DAudioState extends State<Page3DAudio> {
     currentSound = ret.sound;
 
     /// play it
-    final playRet = await SoLoud.instance.play3d(currentSound!, 0, 0, 0);
+    final playRet = await SoLoud.instance.play3d(
+      currentSound!,
+      0,
+      0,
+      0,
+      looping: true,
+    );
     if (playRet.error != PlayerErrors.noError) return;
 
-    SoLoud.instance.setLooping(playRet.newHandle, true);
     SoLoud.instance.set3dSourceMinMaxDistance(
       playRet.newHandle,
       50,
@@ -176,7 +180,7 @@ class _Audio3DWidgetState extends State<Audio3DWidget>
   double posZ = 0;
   double velX = 0;
   double velY = 0;
-  double animVel = 2;
+  double animVel = 1;
   double angle = 0;
   Offset center = Offset.zero;
   late double dx;
@@ -197,8 +201,8 @@ class _Audio3DWidgetState extends State<Audio3DWidget>
 
   void _tick(Duration elapsed) {
     if (widget.spinAround) {
-      angle += pi / animVel / 50;
       final circleRadius = Offset(posX - center.dx, posY - center.dy).distance;
+      angle += pi / (animVel / (circleRadius / widget.width)) / 50;
       posX = circleRadius * cos(angle);
       posY = circleRadius * sin(angle);
     } else {