+++ lib/src/enums.dart
@@ -46,7 +46,8 @@ enum CaptureErrors {
   String toString() => 'CaptureErrors.$name ($_asSentence)';
 }
 
-/// Possible player errors
+/// Possible player errors.
+/// New values must be enumerated at the bottom
 enum PlayerErrors {
   /// No error
   noError,
@@ -84,6 +85,12 @@ enum PlayerErrors {
 
   /// Player not initialized
   backendNotInited,
+
+  /// Filter not found
+  filterNotFound,
+
+  /// asking for wave and FFT is not enabled
+  visualizationNotEnabled,
 
   /// Audio isolate already started
   @Deprecated('Use multipleInitialization instead')