+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 2.0.0-pre.X
+- added `looping` and `loopingStartAt` properties to `SoLoud.play()` and `SoLoud.play3d()`.
+- added `SoLoud.getLooping()` to retrieve the looping state of a sound.
+- added `SoLoud.getLoopPoint()` and `SoLoud.setLoopPoint()` to get and set the looping start position of a sound.
+
 #### 2.0.0-pre.0 (11 Mar 2024)
 - added `bool SoLoud.getVisualizationEnabled()` to get the current state of the visualization.
 - added `mode` property to `SoLoud.loadFile()` and `SoloudTools.loadFrom*` to prevent to load the whole audio data into memory: